Add SystemVerilog conversion system functions (#33)
* Add $signed and $unsigned conversion functions

- Add $signed/$unsigned to system call allow-lists across pipeline
- Implement interpreter handlers to reinterpret signedness
- Implement codegen with static_cast to target type
- Fix Bit<8,true> stream output to cast int8_t to int64_t
- Fix clang-tidy warnings in TimeFormatContext struct

* Add $itor and $rtoi conversion functions

- Add $itor/$rtoi to system call allow-lists across pipeline
- Implement interpreter handlers for int-to-real and real-to-int
- Implement codegen with static_cast to Real and target type
- Add test cases for positive/negative values and truncation

* Add IEEE 754 bit conversion functions

- Add $realtobits/$bitstoreal for 64-bit real <-> bits conversion
- Add $shortrealtobits/$bitstoshortreal for 32-bit shortreal <-> bits
- Implement using std::bit_cast for IEEE 754 compliant reinterpretation
- Add test cases for basic conversions, negative values, and roundtrips

+++ include/lyra/sdk/bit.hpp
@@ -383,7 +383,8 @@ class Bit {
   // Stream output
   friend auto operator<<(std::ostream& os, Bit b) -> std::ostream& {
     if constexpr (Signed) {
-      return os << b.SignedValue();
+      // Cast to largest type to avoid char interpretation for int8_t
+      return os << static_cast<int64_t>(b.SignedValue());
     } else {
       // Cast to largest type to avoid char interpretation for uint8_t
       return os << static_cast<uint64_t>(b.value_);
